+++ packages/terra-framework-docs/CHANGELOG.md
@@ -6,6 +6,7 @@
   * Enabled `close` button in examples.
   * Modified Button in `PopupStandard` example.
   * Removed `ariaLabel` from `PopupStandard` example.
+  * Updated `terra-slide-panel` example to use `terra-buttons`.
 
 ## 1.86.0 - (May 8, 2024)