+++ packages/terra-framework-docs/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 1.87.0 - (May 14, 2024)
+
 * Changed
   * Updated `terra-slide-panel` example to use `terra-buttons`.